Requested changes from #11
