+++ tests/DebounceTest.php
@@ -14,7 +14,6 @@ use Mpbarlow\LaravelQueueDebouncer\Support\SerializingCacheKeyProvider;
 use Mpbarlow\LaravelQueueDebouncer\Tests\Support\CustomCacheKeyProvider;
 use Mpbarlow\LaravelQueueDebouncer\Tests\Support\CustomUniqueIdentifierProvider;
 use Mpbarlow\LaravelQueueDebouncer\Tests\Support\DummyJob;
-use Mpbarlow\LaravelQueueDebouncer\Tests\Support\DummyJobWithArgsAndTrait;
 use ReflectionFunction;
 
 use function array_shift;
@@ -47,24 +46,6 @@ class DebounceTest extends TestCase
         }, PHP_INT_MAX);
     }
     
-    /**
-     * @test
-     */
-    public function it_uses_the_new_trait_to_dispatch()
-    {
-        Bus::fake();
-            
-        $this->mock(Debouncer::class, function ($mock) {
-            $mock
-                ->shouldReceive('debounce')
-                ->once();
-        });
-            
-        DummyJobWithArgsAndTrait::debounce('test', now()->addMinutes(1));
-        
-        //Bus::assertDispatched(DummyJobWithArgsAndTrait::class);
-    }
-
     /**
      * @test
      * @dataProvider jobProvider
